Clean up and simplify cut/copy/paste instructions (#40)
As it was, "Use X or C appropriately for cutting and copying" was not a clear way to communicate, especially when someone probably just wants to skim. The reader probably already knows to use "X" appropriately, they probably don't know to use CTRL + SHIFT + X, which is the whole point of this post. This makes it simpler.

This also removes a claim that it's simple (all things being equal, not mentioning how simple something is makes it simpler by not mentioning simplicity). The congratulations at the end could be seen as patronizing (I took it that way). Also, remove the "If you're new to the Terminal" part because copy-pasting is a nice feature for everyone!